+++ src/features/club-create/hooks/useCreateBookClub.ts
@@ -2,7 +2,7 @@
 
 import { useState } from 'react';
 import { BookClubForm } from '../types';
-import { useBookClubCreateMutation } from '@/features/react-query/book-club';
+import { useBookClubCreateMutation } from '@/api/react-query/book-club';
 
 export const useCreateBookClub = () => {
   const [error, setError] = useState<string | null>(null);
